# core, txpool: less allocations when handling transactions