```
BenchmarkPendingDemotion100-16       	 6804576	       173 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion1000-16      	 6309848	       174 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion10000-16     	 7006021	       176 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion100-16       	358416355	         3.38 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion1000-16      	365451259	         3.37 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion10000-16     	377171361	         3.33 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolBatchInsert100-16       	      38	  27382746 ns/op	  607269 B/op	    8177 allocs/op
BenchmarkPoolBatchInsert1000-16      	       5	 259864088 ns/op	 5861854 B/op	   83225 allocs/op
BenchmarkPoolBatchInsert10000-16     	       1	2085832352 ns/op	51360392 B/op	  690243 allocs/op
BenchmarkPoolBatchInsert100000-16    	       1	17361944530 ns/op	445913240 B/op	 5550197 allocs/op
```
Master
```
BenchmarkPendingDemotion100-16      	 5902473	       183 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion1000-16     	 6397833	       178 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion10000-16    	 6575794	       176 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion100-16      	389451097	         2.96 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion1000-16     	371748506	         2.96 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion10000-16    	404261654	         3.01 ns/op	       0 B/op	       0 allocs/op
BenchmarkPoolBatchInsert100-16      	      39	  27335098 ns/op	  633578 B/op	    9082 allocs/op
BenchmarkPoolBatchInsert1000-16     	       6	 216564796 ns/op	 6051392 B/op	   94093 allocs/op
BenchmarkPoolBatchInsert10000-16    	       1	2149670407 ns/op	53134952 B/op	  750930 allocs/op
```