@holiman 2020-06-17
I'm golfing away
```
BenchmarkPoolBatchInsert100-6                 67          19192397 ns/op          373139 B/op       5148 allocs/op
BenchmarkPoolBatchInsert1000-6                 6         190592352 ns/op         3704962 B/op      51542 allocs/op
BenchmarkPoolBatchInsert10000-6                1        2091724254 ns/op        33357816 B/op     498078 allocs/op
```