@MariusVanDerWijden 2020-06-17
Now down to
```
BenchmarkPoolBatchInsert100-16    	      37	  27332816 ns/op	  379164 B/op	    5002 allocs/op
BenchmarkPoolBatchInsert1000-16    	       6	 217340430 ns/op	 3701454 B/op	   50551 allocs/op
BenchmarkPoolBatchInsert10000-16    	       1	2080770716 ns/op	33212728 B/op	  478450 allocs/op
```