@MariusVanDerWijden 2020-06-17
With the new test cases we spot the big memory leak: 
```
BenchmarkPoolBatchInsert100-16                        43          27457989 ns/op          378687 B/op       5002 allocs/op
BenchmarkPoolBatchInsert1000-16                        6         203928329 ns/op         3698616 B/op      50541 allocs/op
BenchmarkPoolBatchInsert10000-16                       1        1967694228 ns/op        33202688 B/op     478333 allocs/op
BenchmarkPoolBatchLocalInsert100-16                   57          69215142 ns/op         2583880 B/op       5124 allocs/op
BenchmarkPoolBatchLocalInsert1000-16                   6         880639775 ns/op        35409178 B/op      51022 allocs/op
BenchmarkPoolBatchLocalInsert10000-16                  1        32001544662 ns/op       1465866464 B/op   554108 allocs/op
```
```
BenchmarkPoolBatchInsert100-16                        66          27111004 ns/op          614724 B/op       9222 allocs/op
BenchmarkPoolBatchInsert1000-16                        6         206998251 ns/op         6050997 B/op      94097 allocs/op
BenchmarkPoolBatchInsert10000-16                       1        2068979714 ns/op        53141072 B/op     750973 allocs/op
BenchmarkPoolBatchLocalInsert100-16                   52          42339111 ns/op         3612176 B/op     338515 allocs/op
BenchmarkPoolBatchLocalInsert1000-16                   5         200485642 ns/op         6020937 B/op      89963 allocs/op
BenchmarkPoolBatchLocalInsert10000-16                  1        109329517576 ns/op      14425409792 B/op        1617470241 allocs/op
```