@MariusVanDerWijden 2020-06-18
```
> benchcmp master tx_reduceallocsv2
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                 old ns/op        new ns/op      delta
BenchmarkPoolBatchInsert100-16            26964301         27592412       +2.33%
BenchmarkPoolBatchInsert1000-16           218805225        224426000      +2.57%
BenchmarkPoolBatchInsert10000-16          2087232704       2088864320     +0.08%
BenchmarkPoolBatchLocalInsert100-16       21966964         24972428       +13.68%
BenchmarkPoolBatchLocalInsert1000-16      2551680926       239331292      -90.62%
BenchmarkPoolBatchLocalInsert10000-16     110797073688     3842275783     -96.53%

benchmark                                 old allocs     new allocs     delta
BenchmarkPoolBatchInsert100-16            9090           5003           -44.96%
BenchmarkPoolBatchInsert1000-16           94089          50540          -46.28%
BenchmarkPoolBatchInsert10000-16          750991         478352         -36.30%
BenchmarkPoolBatchLocalInsert100-16       9068           5000           -44.86%
BenchmarkPoolBatchLocalInsert1000-16      34290177       49235          -99.86%
BenchmarkPoolBatchLocalInsert10000-16     1617469794     491134         -99.97%

benchmark                                 old bytes       new bytes     delta
BenchmarkPoolBatchInsert100-16            633746          377866        -40.38%
BenchmarkPoolBatchInsert1000-16           6051782         3697901       -38.90%
BenchmarkPoolBatchInsert10000-16          53142120        33209024      -37.51%
BenchmarkPoolBatchLocalInsert100-16       644781          391006        -39.36%
BenchmarkPoolBatchLocalInsert1000-16      311539462       3580522       -98.85%
BenchmarkPoolBatchLocalInsert10000-16     14425394808     36389416      -99.75%
```
most additional bytes spent is due to the benchmark being dependent on `b.N` e.g. 
`BenchmarkPoolBatchLocalInsert100-16` runs on master with `b.N = 51` and on this branch with `b.N = 63` which means this branch inserts 1200 transactions more (12 * 100). 

The overhead of linearly searching the pool can be seen on the global batch insert benchmarks and should be around ~1-2%, since we reduce the allocations by ~50% this small regression should be negligible since we reduce the need for the garbage collector to kick in drastically.