@holiman 2020-06-18
This is now up for testing on prod: https://geth-bench.ethdevops.io/d/QC1Arp5Wk/single-geth?orgId=1&var-host=bootnode-aws-eu-west-1-001&var-percentile=50&from=1592490879131&to=now -- graph from 30 minutes before updating to this PR