@holiman 2020-06-21
Some comparisons with other bootnodes, 7 days. This PR (on `aws-eu-west`) was swapped in ~3 days ago, and is shown in green. This is the alloc-count
![Screenshot_2020-06-21 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/85235291-1c939100-b414-11ea-8837-e6ee85e64d26.png)
![Screenshot_2020-06-21 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/85235292-1d2c2780-b414-11ea-9e26-60ca402fb976.png)

This is the alloc-count for all bootnodes, last 12 hours, stacked. 
![Screenshot_2020-06-21 Multi Geth - Grafana](https://user-images.githubusercontent.com/142290/85235296-31702480-b414-11ea-902d-298cde39dc3c.png)

This PR is the third from the bottom, light turqoise. 
And finally, the legends showing the average and current, last `12` hours: 

![Screenshot_2020-06-21 Multi Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/85235379-bbb88880-b414-11ea-88f3-514ca4ecf634.png)
Same for 2 days: 

![Screenshot_2020-06-21 Multi Geth - Grafana(2)](https://user-images.githubusercontent.com/142290/85235400-e276bf00-b414-11ea-9ce9-6a04c68ad357.png)