@fjl 2020-07-09
Reverted in https://github.com/ethereum/go-ethereum/commit/bcb308745010675671991522ad2a9e811938d7fb because this PR handles account balances as uint64, which is not OK without overflow checks.